Moved nonreciprocal nonsubwavelength logic to NonReciprocalSWP1D
